css for frontend and split ConversationPanel
