+++ message_app_client/src/components/conversation_detail/ConversationHeader.js
@@ -3,7 +3,7 @@ import React from 'react';
 const ConversationHeader = (props) => {
     const { otherParticipantName } = props;
 
-    return <div>You are in a conversation with {otherParticipantName}</div>
+    return <div className='conversation-header'>You are in a conversation with {otherParticipantName}</div>
 };
 
 export default ConversationHeader;
