+++ message_app_client/src/components/conversation_detail/ConversationInput.js
@@ -4,10 +4,10 @@ import TextInput from './TextInput';
 
 const ConversationInput = (props) => {
     return (
-        <>
+        <div className='conversation-input'>
             <ImageInput />
             <TextInput />
-        </>
+        </div>
     )
 };
