+++ message_app_client/src/components/conversation_detail/ImageInput.js
@@ -2,9 +2,9 @@ import React from 'react';
 
 const ImageInput = (props) => {
     return (
-        <>
+        <div className='image-input'>
             <input type="file" id="send-image" accept=".gif,.jpg,.jpeg,.png"/><br></br>
-        </>
+        </div>
     )
 };
