+++ message_app_client/src/components/conversation_detail/Message.js
@@ -4,9 +4,10 @@ import MessageContent from './MessageContent';
 
 const Messages = (props) => {
     const {id, senderName, content, createdAt, username} = props;
+    const messageCSS = "message" + (senderName === username ? "-yourself" : "-otherParticipant") 
 
     return (
-        <>
+        <div className={messageCSS}>
             <MessageHeader
                 senderName={senderName}
                 sentTime={createdAt}
@@ -16,9 +17,8 @@ const Messages = (props) => {
                 content={content}
                 isImage={false}
             />
-        </>
+        </div>
     )
 };
 
 export default Messages;
-
