+++ message_app_client/src/components/conversation_detail/MessageContent.js
@@ -4,13 +4,13 @@ const MessageContent = (props) => {
     const {content, isImage} = props;
 
     return (
-        <>
+        <div className="message-content">
             {isImage ? (
                 <div></div>  
             ) : (
                 <div>{content}</div>
             )}
-        </>
+        </div>
     );
 };
