+++ message_app_client/src/components/conversation_panel/ConversationContainer.js
@@ -1,0 +1,18 @@
+import React from 'react';
+import ConversationCard from './ConversationCard';
+
+function ConversationContainer(props){
+    // props.conversations: Array of objects containing conversation's data
+
+    const renderConversation = (conversation) => {
+        return <ConversationCard {...conversation} key={conversation.conversationID}/>
+    }
+
+    return (
+        <div id="conversation-container" className='conversation-container'>CONVERSATIONS
+            {props.conversations.map(conversation => renderConversation(conversation))}
+        </div>
+    )
+}
+
+export default ConversationContainer;
