+++ message_app_client/src/components/conversation_panel/ConversationPanel.css
@@ -1,0 +1,38 @@
+.conversation-panel {
+  flex-grow: 1;
+
+  display: flex;
+  flex-direction: column;
+
+  text-align: center;
+  background-color: white;
+  border: 0.5px solid rgb(137, 137, 155);
+  border-radius: 15px;
+  margin: 2%;
+  margin-top: 5%;
+}
+
+.conversation-container {
+  display: flex;
+  flex-direction: column;
+  overflow-y: scroll;
+  margin: 2%;
+}
+
+.conversation-card {
+  display: flex;
+  flex-direction: column;
+
+  text-align: left;
+  color: black;
+  background-color: #92a8d1;
+  border: 0.5px solid rgb(137, 137, 155);
+  border-radius: 10px;
+  padding: 4%;
+  margin: 2%;
+}
+
+.last-message {
+  display: flex;
+  color: black;
+}
